Feature/feed vector

+++ src/main/java/ar/edu/unq/eperdemic/services/FeedService.kt
@@ -4,6 +4,6 @@ import ar.edu.unq.eperdemic.modelo.Evento
 
 interface FeedService {
     fun feedPatogeno(tipoDePatogeno: String): List<Evento>
-    fun feedVector(vectorId: Int): List<Evento>
+    fun feedVector(vectorId: Long): List<Evento>
     fun feedUbicacion(nombreDeUbicacion: String): List<Evento>
 }
